👾 InputNumber 优化 交互前生效、优化 Pagination UI、修复 Dropdown Trigger.Hover 多级无法停留、修复一些图缺损

+++ src/AntdUI/Controls/Carousel.cs
@@ -406,7 +406,7 @@ namespace AntdUI
                         bmpcode = select_range.i;
                         bmp = PaintBmp(select_range, rect, _radius);
                     }
-                    g.DrawImage(bmp, rect.X - AnimationChangeValue, rect.Y, bmp.Width, bmp.Height);
+                    g.DrawImage(bmp, (int)(rect.X - AnimationChangeValue), rect.Y, bmp.Width, bmp.Height);
                 }
                 else g.PaintImg(rect, image, imageFit, _radius, round);
             }
